+++ ChangeLog.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for morph targets / blend shapes (#8)
 - Support for animated morph targets / blend shapes
-- Limited support for sparse accessors (float vertex positions only for now)
+- Limited support for sparse accessors (vertex positions only for now)
 - Safe build option for more robust loading (`GLTFAST_SAFE` scripting define)
 - Burst as dependency
 ### Changed